+++ TabletopTweaks/NewContent/ContentAdder.cs
@@ -130,6 +130,7 @@ namespace TabletopTweaks.NewContent {
 
                 MythicAbilities.FavoriteMetamagicPersistent.AddFavoriteMetamagicPersistent();
                 MythicAbilities.FavoriteMetamagicSelective.AddFavoriteMetamagicSelective();
+                MythicAbilities.DimensionalRetribution.AddDimensionalRetribution();
 
                 Feats.ShatterDefenses.AddNewShatterDefenseBlueprints();
                 Feats.MagicalAptitude.AddMagicalAptitude();